upd hasValidReceiverTypeKDoc()
