+++ lib/src/main/kotlin/com/lemonappdev/konsist/api/ext/provider/KoReceiverTypeProviderExt.kt
@@ -1,5 +1,7 @@
 package com.lemonappdev.konsist.api.ext.provider
 
+import com.lemonappdev.konsist.api.KoKDocTag
+import com.lemonappdev.konsist.api.provider.KoKDocProvider
 import com.lemonappdev.konsist.api.provider.KoReceiverTypeProvider
 
 /**
@@ -8,3 +10,14 @@ import com.lemonappdev.konsist.api.provider.KoReceiverTypeProvider
  * @return `true` if the declaration has receiver type with the specified KClass name, `false` otherwise.
  */
 inline fun <reified T> KoReceiverTypeProvider.hasReceiverTypeOf(): Boolean = T::class.simpleName == receiverType?.name
+
+/**
+ * Whether declaration has a valid KDoc with a RECEIVER tag.
+ *
+ * @return `true` if the declaration has a valid KDoc with the RECEIVER tag, `false` otherwise.
+ */
+fun <T : KoReceiverTypeProvider> T.hasValidReceiverTypeKDoc(): Boolean = if (receiverType != null) {
+    (this as? KoKDocProvider)?.kDoc?.hasTags(KoKDocTag.RECEIVER) == true
+} else {
+    true
+}
