+++ lib/src/main/kotlin/com/lemonappdev/konsist/core/provider/KoReceiverTypeProviderCore.kt
@@ -10,6 +10,7 @@ import org.jetbrains.kotlin.psi.psiUtil.isExtensionDeclaration
 internal interface KoReceiverTypeProviderCore :
     KoReceiverTypeProvider,
     KoContainingDeclarationProviderCore,
+    KoKDocProviderCore,
     KoBaseProviderCore {
     val ktCallableDeclaration: KtCallableDeclaration
